[Task-008] - Student 4 - Mandatory #90
Dashboard added.

Also added a service and a repository for authenticated users, in order
to make calls to an API and, with this, we can have better dashboard
with the same currency.

Also say that the data of the API is saved in the DDBB, and it will be
actualized each hour if requested.