[Task-018] - Student 4 - Suplementary
List mine and list all published, separeted. Now everybody can list the
sponsorships.

Also added the buttons to get to the diferent options.

Fix the collection of projects for a sponsorship, for getting all the
published and not the ones the sponsor use.

Changed also the sponsorship.csv, because there where some sponsorships
connected to a project not published, so there where errors in there.